Cast aux to match dtype

+++ fms/modules/attention.py
@@ -615,7 +615,7 @@ class MultiHeadAttention(nn.Module):
             )  # b h l d
             attn = attn.transpose(1,2).contiguous()  # b l h d
             affs = mask.view(batch_size, self.kvheads, -1, mask.size(-2), mask.size(-1))[:,:,0,-1].exp()  # b h l
-            aux = affs.gt(.001).to(dtype=affs.dtype).sub(affs.detach()).add(affs).view(batch_size, -1).mean(-1)
+            aux = affs.gt(.001).to(dtype=affs.dtype).sub(affs.detach()).add(affs).view(batch_size, -1).mean(-1).to(dtype=queries.dtype)
 
             # c = 512
             # b = batch_size
